pyproject.toml
[tool.poetry]
name = "hw24-1"
version = "0.1.0"
description = ""
authors = ["Eduard Slobodyanik <slobodyanik.ed@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
black = "^24.10.0"
isort = "^5.13.2"
django = "4.2.2"
crispy-bootstrap4 = "^2024.10"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.48"
python-dotenv = "^1.0.1"
redis = "^5.2.0"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.29.0"
coverage = "^7.6.4"
drf-yasg = "^1.21.8"
stripe = "^11.2.0"
forex-python = "^1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
